# Split timeseries by data types in Metrics Protobuf definitions