Previously we had Point message which was a value of oneof the data types. This is
unnecessary flexibility because points in the same timeseries cannot be of different
data type. This also costed performance.

Now we have separate timeseries message definitions for each data type and the
timeseries used is defined by the oneof entry in the Metric message.

This change is stacked on top of https://github.com/open-telemetry/opentelemetry-proto/pull/33

Simple benchmark in Go demonstrates the following improvement of encoding and decoding
compared to the baseline state:

```
===== Encoded sizes
Encoding                       Uncompressed  Improved        Compressed  Improved
Baseline/MetricOne              24200 bytes  [1.000], gziped 1813 bytes  [1.000]
Proposed/MetricOne              19150 bytes  [1.264], gziped 1586 bytes  [1.143]

Encoding                       Uncompressed  Improved        Compressed  Improved
Baseline/MetricSeries           56022 bytes  [1.000], gziped 6691 bytes  [1.000]
Proposed/MetricSeries           43115 bytes  [1.299], gziped 6328 bytes  [1.057]

goos: darwin
goarch: amd64
pkg: github.com/tigrannajaryan/exp-otelproto/encodings
BenchmarkEncode/Baseline/MetricOne-8         	      26	 212691770 ns/op	27376010 B/op	  351000 allocs/op
BenchmarkEncode/Proposed/MetricOne-8         	      50	 122404576 ns/op	20880001 B/op	   51000 allocs/op

BenchmarkEncode/Baseline/MetricSeries-8      	       8	 653270997 ns/op	71344004 B/op	 1751000 allocs/op
BenchmarkEncode/Proposed/MetricSeries-8      	      19	 313939544 ns/op	51152000 B/op	  251000 allocs/op

BenchmarkDecode/Baseline/MetricOne-8         	      14	 383343126 ns/op	186296037 B/op	 5274000 allocs/op
BenchmarkDecode/Proposed/MetricOne-8         	      24	 250832929 ns/op	156696050 B/op	 4274000 allocs/op

BenchmarkDecode/Baseline/MetricSeries-8      	       5	1031024029 ns/op	455096041 B/op	12174000 allocs/op
BenchmarkDecode/Proposed/MetricSeries-8      	      10	 567579261 ns/op	339096060 B/op	 8374000 allocs/op
```

It is 30-50% faster and is 20-25% smaller on the wire and in memory.

Benchmarks encode and decode 500 batches of 2 metrics: one int64 Gauge with 5 time series
and one Histogram of doubles with 1 time series and single bucket. Each time series for
both metrics contains either 1 data point (MetricOne) or 5 data points (MetricSeries).
Both metrics have 2 labels.

Benchmark source code is available at:
https://github.com/tigrannajaryan/exp-otelproto/blob/master/encodings/encoding_test.go